Add package uninstall endpoint
Add a new endpoint that can be used to uninstall a package.

The implementation follows the current behavior of the dcos-cli, namely:
1. list all apps in marathon, filtered to those apps with package names
matching the supplied `name`
2. If more than 1 app and `all` is set to `true` proceed, otherwise
return an error
3. for each app to remove do a force delete from marathon, if a
framework name label is present query mesos for a list of frameworks,
filter where framework name matches the value of the label. If more than
one framework return an error otherwise proceed to tear down the
framework.
4. Upon success return the deails of all apps that were removed