Remove CosmosResult type
This is being done after a discussion regarding project work velocity. For now the additional overhead of dealing with CosmosResult and xor is adding a fair amount of additional work in terms of getting the compiler to carry things through the way we want/expect.

The idea of CosmosResult may resurface in the future but for now it's being ripped out.