Add new class EndpointHandler[Req, Rep]
Add new class EndpointHandler[Req, Rep] that provides enforcement of the
Content-Type and Accept type headers. Each EndpointHandler defines the
type of it's input and output and contains a corresponding
RequestReader[Req] that will ensure the headers are set correspondingly
for requests sent to the endpoint.

Responses will also have a custom Content-Type set corresponding to the
type of the response. This can be used to correlate responses to schemas
as well as informing client how the data should be interprited.

For any "error response" (http 4xx or 5xx) the Content-Type will
correspond with the CosmosError type.