Remove top level errors from error response.
Top level errors property has been removed from ErrorResponse in favor of
using the data property.

data is now an Option[JsonBody] instead of a Json with a default value.
This decision was made to ensure that we don't end up with data that is
unhelpful simple types (string, number, boolean). If an error needs to express
multiple sub-errors a nested object can be created that contains each of the
errors.