Stop expecting warnings in JSON schema validation test for Chronos
The config.json for Chronos in the cli-test-4 branch of the Universe
repository was recently updated to no longer generate warnings when
processed by our JSON schema validator. This had the side effect of
causing a test to fail that expected those warnings to be returned
in the package install response for Chronos.