Implement list-sources, add-source, and delete-source endpoints
Introduces a PackageSourcesStore trait which manages the storage
and retrieval of the package sources list; this has a concrete
implementation which uses ZooKeeper as the backing store.

There are a number of implementation details that will need to be
refined in later changes, such as ensuring data integrity for
concurrent updates, and integrating the sources list with the
other endpoints so that they actually make use of multiple sources.