Implement universe cache lifecycle
This change includes the following improvements:

1. We automatically update the local cache on every request if the cache
is older than one minutes. This is currently not configurable. We will
make it configurable once we add support for changing configuration to
Cosmos.

2. Change the constructor for the universe cache so that it return a
sync value and not an async value. We have moved all of the async
operation fo the public methods of PackageCache.

3. Better handle the creation of the local universe directory. This
commit first creates a temporary directory and a new symbolic link in
the same local cache directory and atomically moves the new symbolic
link to a symbolic link named after base64 encoded url for the universe
repo/source.

4. Updating the local cache is serialized using an AsyncMutex. This
guarrantes that we only do at most one HTTP against Universe per minute.

5. Added Apache commos-io for their implementation of recurse delete of
directories.