Create ZooKeeperUri class for validation of ZooKeeper configuration
An instance of the class can only be constructed via successful
parsing of a string in the format:

zk://host1:port1,host2:port2,host3:port3/path/to/chroot/znode

The port numbers are optional and the number of hosts is flexible,
but there must be at least one. There must also be at least one path
component.

The class is currently only used when parsing command-line flags.