Improve error reporting for repository Zip download (#302)
Replace the `CosmosError` subclass `InvalidRepositoryUri` with
`RepositoryUriSyntax` and `RepositoryUriConnection`. The JSON
encodings of these new types include the message from the
`Throwable` cause of the error, to further assist with diagnosis.

To facilitate unit testing, the `UniversePackageCache` method
`updateUniverseCache` was moved to the companion object, allowing
it to be tested without creating an instance of a cache. This allowed
most of the tests to be written without use of the filesystem.

Fixes #299.