Refactor ListHandler to be readable and testable
* Split the apply method into sensible steps each yielding a future
	* get applications
	* get application to repository associations
	* Given the repo, get the installation of each application

These methods are bulk operations that are at this point doing a lot of
calls that yield lots of futures and then aggregate them into one big future.
In the future these bulk methods could be pushed to the receivers so that
they are less costly.

The organisation of the code into these three methods also makes the
code more testable, since given mocks for adminRouter and repositories
we could unit test them independently.