Filling the release version field on conversion
When converting from a Metadata to a V3Package we need to write the
current time in UTC as the number of seconds from the Unix Epoch. This
means that we need to convert ReleaseVersion to use a Long instead of a
Int. This is okay as JSON numberse are magical.