Store only the user provided options (#602)
Change the options JSON object that we store in Marathon so that we only
store the user provided options. If the user doesn't provide any option
that is equivalent to the empty JSON objct. E.g. "{}".

The possible values for the DCOS_PACKAGE_OPTIONS are:
1. Missing - the application was deployed with an old
Cosmos so we don't have the user provided options.
2. Empty JSON object - the user didn't provide any options
3. JSON object - contains the user provided options