[DCOS-44443] Remove the deprecated DCOS_PACKAGE_METADATA label (#760)
Address [DCOS-44443](https://jira.mesosphere.com/browse/DCOS-44443)

We have deprecated the DCOS_PACKAGE_METADATA label from 1.11 onwards. DC/OS Frontend was the only component that was using this label. They switched to use `/service/describe` instead of this label (And that was needed for `package-registry` to work anyway). Since this label is deprecated, we can remove the code to encode/decode this label.