make it easier to run integration tests locally
you now only need to provide that dcosUri systemProperty in order to
have your integration tests running. we can also remove a step from our
TeamCity build now.

yes, we implicitly add a dependency on jq here in some random bash
script - but we do that on TC anyways and will shortly add a Dockerfile
to make sure we don't shoot ourselves in the foot with that.

until then it's way more convenient to run integration tests.

i thought of using an ENV-var to provide a CLUSTER_URL instead of a
system property - but that seems to need adaptions to some production
code i don't want to touch that early in my scala career ;)